fix: speed calculation, DAO precision, AppTheme, and FakeLocationSource
- PointsWorker: convert timeDelta from ms to seconds before computing speed
- PointsAwardedDao: change SUM return types from Int to Double to preserve precision
- DefaultGameRepository: map Double->Int at boundary for UI consumers
- ComposeDetailActivity: wrap setContent in AppTheme for proper theming
- FakeLocationSource: add createDomainLocation() for pure JVM tests

+++ points/src/main/java/com/adsamcik/tracker/points/database/PointsAwardedDao.kt
@@ -15,13 +15,13 @@ interface PointsAwardedDao : BaseDao<PointsAwarded> {
 	 * Returns number of points earned between two time intervals.
 	 */
 	@Query("SELECT COALESCE(SUM(value), 0) FROM points_awarded WHERE time >= :from AND time <= :to")
-	fun countBetween(from: Long, to: Long): Int
+	fun countBetween(from: Long, to: Long): Double
 
 	/**
 	 * Returns number of points earned between two time intervals as a Flow.
 	 */
 	@Query("SELECT COALESCE(SUM(value), 0) FROM points_awarded WHERE time >= :from AND time <= :to")
-	fun countBetweenFlow(from: Long, to: Long): Flow<Int>
+	fun countBetweenFlow(from: Long, to: Long): Flow<Double>
 
 	@Query("SELECT EXISTS(SELECT 1 FROM points_awarded WHERE time = :time AND source = :source LIMIT 1)")
 	fun hasAwardAt(time: Long, source: String): Boolean
